+++ userspace/lib/include/signal.h
@@ -1,6 +1,8 @@
 #ifndef SRVROS_POSIX_SIGNAL_H
 #define SRVROS_POSIX_SIGNAL_H
 
+#include <sys/types.h>
+
 typedef int sig_atomic_t;
 typedef unsigned long sigset_t;
 typedef void (*sighandler_t)(int);
@@ -12,6 +14,8 @@ typedef void (*sighandler_t)(int);
 #define SIG_ERR ((sighandler_t)-1)
 
 sighandler_t signal(int signum, sighandler_t handler);
+int kill(pid_t pid, int sig);
+int raise(int sig);
 int sigemptyset(sigset_t *set);
 int sigfillset(sigset_t *set);
 int sigaddset(sigset_t *set, int signum);